Generics and Java 7 syntax in plugins package

+++ src/java/org/apache/ivy/plugins/repository/RepositoryCopyProgressListener.java
@@ -31,7 +31,7 @@ public class RepositoryCopyProgressListener implements CopyProgressListener {
 
     public void start(CopyProgressEvent evt) {
         if (totalLength != null) {
-            repository.fireTransferStarted(totalLength.longValue());
+            repository.fireTransferStarted(totalLength);
         } else {
             repository.fireTransferStarted();
         }
